CORE-593 + CORE-590 extends and refines anomalies for index and contact
The temperature anomaly is reactivate for index cases.

Now there are two thresholds for body temperature to differentiate index
and contact.

In addition to the first characteristic symptom, symptoms can now also
be marked as conspicuous separately for index cases and contact persons
(new entity properties and database columns). With this also introduces
a new anomaly that can be considered as a higher escalation level to the
first characteristic symptom. For index cases only this anomaly is
relevant.